+++ window.go
@@ -34,9 +34,9 @@ func (w *Window) Size() (int, int) {
 	return w.glfwWin.GetSize()
 }
 
-func (w *Window) Aspect() float64 {
+func (w *Window) Aspect() float32 {
 	width, height := w.Size()
-	return float64(width) / float64(height)
+	return float32(width) / float32(height)
 }
 
 func (w *Window) updateGraphics() {
